fix: prevent annotations from disappearing due to their being out-of-order
This change updates the merged list sort key to use `pageno` instead of a timestamp.

When arriving on a new page, KOReader appears to scan through annotations looking for ones for that page.

From the behavior I observed through experimentation, it appears that it
uses an algorithm like this:

- Read annotations from the annotations table in order until a matching `pageno` is found.
- Continue reading and collecting annotations until `pageno` is greater than the current page.
- Stop reading annotations.

Due to this optimization, if we order annotations by timestamp before
writing them back out, we can create a condition where an annotation
added at a later date to a location in the book that is on an earlier
page than an existing annotation, that new annotation will disappear
from KOReader.

Annotations may appear in the KOReader sdr lua in non-sequential order
within the subset of entries that have the same `pageno` value as long
as all `pageno` groupings are in the correct order. So `pageno` ordering
alone is sufficient.

Fixes #6